+++ src/main/CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Ender Converter ignoring other Converters on signal validation
+- Ender Converter not notifying other Ender Converters of change
 
 ## [1.0.2] - 2016-01-30
 ### Fixed